Fix compilation on x86_64-uwp-windows-msvc.
