+++ src/imp/libc/io/windows_syscalls.rs
@@ -1,10 +1,12 @@
 //! Windows system calls in the `io` module.
 
-use super::super::conv::{borrowed_fd, ret};
+use super::super::conv::{borrowed_fd, ret, ret_c_int};
 use super::super::fd::LibcFd;
 use super::c;
 use crate::fd::{BorrowedFd, RawFd};
 use crate::io;
+use crate::io::PollFd;
+use core::convert::TryInto;
 
 pub(crate) unsafe fn close(raw_fd: RawFd) {
     let _ = c::close(raw_fd as LibcFd);
@@ -15,4 +17,14 @@ pub(crate) fn ioctl_fionbio(fd: BorrowedFd<'_>, value: bool) -> io::Result<()> {
         let mut data = value as c::c_uint;
         ret(c::ioctl(borrowed_fd(fd), c::FIONBIO, &mut data))
     }
+}
+
+pub(crate) fn poll(fds: &mut [PollFd<'_>], timeout: c::c_int) -> io::Result<usize> {
+    let nfds = fds
+        .len()
+        .try_into()
+        .map_err(|_convert_err| io::Error::INVAL)?;
+
+    ret_c_int(unsafe { c::poll(fds.as_mut_ptr().cast(), nfds, timeout) })
+        .map(|nready| nready as usize)
 }
